Report progress based on downloaded content
When swupd reports progress of content it needs, it does so by counting
how many files it needs to download (fullfiles or packs) and how many it
has already downloaded. This gives a rough estimate of what is the
progress of the overall download, but it can also be very misleading
since some files may be very different in size compared to others. This
is specially true when talking about packs, one pack could be a couple
of megabytes big while another one could be a few hundred megabytes.

This commit adds a curl callback that can be used to report download
progress periodically based on how many bytes have been downloaded vs
how many bytes have to be downloaded in total, giving the ability to report
progress accurately.

+++ src/lib/progress.c
@@ -87,7 +87,7 @@ void progress_complete_step(void)
 	}
 }
 
-void progress_report(unsigned int count, unsigned int max)
+void progress_report(double count, double max)
 {
 	static int last_percentage = -1;
 	static unsigned int last_step = 0;
@@ -96,7 +96,7 @@ void progress_report(unsigned int count, unsigned int max)
 	if (max != 0) {
 
 		/* Only print when the percentage changes, so a maximum of 100 times per run */
-		int percentage = (int)(100 * ((float)count / (float)max));
+		int percentage = (int)(100 * (count / max));
 		if (percentage != last_percentage || step.current != last_step) {
 			if (progress_function) {
 				progress_function(step.description, step.current, step.total, percentage);
